libzfs/mnttab: lift mnttab cache into separate file
Sponsored-by: TrueNAS
