+++ lib/libzfs/Makefile.am
@@ -16,6 +16,7 @@ dist_libzfs_la_SOURCES = \
 	%D%/libzfs_diff.c \
 	%D%/libzfs_import.c \
 	%D%/libzfs_iter.c \
+	%D%/libzfs_mnttab.c \
 	%D%/libzfs_mount.c \
 	%D%/libzfs_pool.c \
 	%D%/libzfs_share.c \
